Changed code in the Shopping cart component and then changed the name on the component to OutputShoppingCart
