+++ src/components/CartIcon/index.tsx
@@ -1,6 +1,6 @@
 import * as React from "react";
 import * as Styles from "./index.styles";
-import UseCartStore from "../ShoppingCart/UseCartStore";
+import UseCartStore from "../OutputShoppingCart/UseCartStore";
 import { Link } from "react-router-dom";
 
 function CartIcon() {
