+++ src/components/OutputShoppingCart/getPricesFromCart.ts
